+++ src/cn/edu/zhku/jsj/dao/FoodDao.java
@@ -2,7 +2,7 @@ package cn.edu.zhku.jsj.dao;
 
 import java.util.List;
 
-import cn.edu.zhku.jsj.daomain.Food;
+import cn.edu.zhku.jsj.domain.Food;
 
 public interface FoodDao {
