+++ src/cn/edu/zhku/jsj/dao/OrderDao.java
@@ -3,7 +3,7 @@ package cn.edu.zhku.jsj.dao;
 import java.util.List;
 import java.util.Map;
 
-import cn.edu.zhku.jsj.daomain.Order;
+import cn.edu.zhku.jsj.domain.Order;
 
 public interface OrderDao {
