+++ src/cn/edu/zhku/jsj/dao/impl/ClothDaoImpl.java
@@ -8,7 +8,7 @@ import java.util.List;
 import javax.management.RuntimeErrorException;
 
 import cn.edu.zhku.jsj.dao.ClothDao;
-import cn.edu.zhku.jsj.daomain.Cloth;
+import cn.edu.zhku.jsj.domain.Cloth;
 import cn.edu.zhku.jsj.web.utils.JdbcUtil;
 import cn.edu.zhku.jsj.web.utils.ResultToBean;
