+++ src/cn/edu/zhku/jsj/dao/impl/FoodDaoImpl.java
@@ -7,8 +7,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import cn.edu.zhku.jsj.dao.FoodDao;
-import cn.edu.zhku.jsj.daomain.Cloth;
-import cn.edu.zhku.jsj.daomain.Food;
+import cn.edu.zhku.jsj.domain.Cloth;
+import cn.edu.zhku.jsj.domain.Food;
 import cn.edu.zhku.jsj.web.utils.JdbcUtil;
 import cn.edu.zhku.jsj.web.utils.ResultToBean;
