+++ src/cn/edu/zhku/jsj/dao/impl/OrderDaoImpl.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Map;
 
 import cn.edu.zhku.jsj.dao.OrderDao;
-import cn.edu.zhku.jsj.daomain.Order;
+import cn.edu.zhku.jsj.domain.Order;
 
 public class OrderDaoImpl implements OrderDao {
 	@Override
