+++ src/cn/edu/zhku/jsj/domain/Cloth.java
@@ -1,17 +1,16 @@
-package cn.edu.zhku.jsj.daomain;
+package cn.edu.zhku.jsj.domain;
 
-public class Book {
-	private int book_id;
+public class Cloth {
+	private int cloth_id;
 	/*private String type;//记录种类
-*/	private String bookname;
-	private String author;
-	private String press;//出版社
+*/	private String clothname;
 	private float price;
+	private String color;  //颜色选择
+	private String size;
 	private String version;//记录用户定义的标签,用","隔开 取数据的时候用<c:forTokens> 标签 
-	private String images; //记住各种书的 图片
-	private String ISBN;
+	private String images;//记住衣服 的各种图片
 	private String description;
-	private int store_id;  //外键 记住是哪家店的商品
+	private int store_id;  //外键 记住是那哪店的商品
 	private int totalnum;   //货品的库存量
 	public int getTotalnum() {
 		return totalnum;
@@ -19,29 +18,18 @@ public class Book {
 	public void setTotalnum(int totalnum) {
 		this.totalnum = totalnum;
 	}
-	public int getBook_id() {
-		return book_id;
+	public int getCloth_id() {
+		return cloth_id;
 	}
-	public void setBook_id(int book_id) {
-		this.book_id = book_id;
+	public void setCloth_id(int cloth_id) {
+		this.cloth_id = cloth_id;
 	}
-	public String getBookname() {
-		return bookname;
+	
+	public String getClothname() {
+		return clothname;
 	}
-	public void setBookname(String bookname) {
-		this.bookname = bookname;
-	}
-	public String getAuthor() {
-		return author;
-	}
-	public void setAuthor(String author) {
-		this.author = author;
-	}
-	public String getPress() {
-		return press;
-	}
-	public void setPress(String press) {
-		this.press = press;
+	public void setClothname(String clothname) {
+		this.clothname = clothname;
 	}
 	public float getPrice() {
 		return price;
@@ -49,24 +37,30 @@ public class Book {
 	public void setPrice(float price) {
 		this.price = price;
 	}
+	public String getColor() {
+		return color;
+	}
+	public void setColor(String color) {
+		this.color = color;
+	}
+	public String getSize() {
+		return size;
+	}
+	public void setSize(String size) {
+		this.size = size;
+	}
 	public String getVersion() {
 		return version;
 	}
 	public void setVersion(String version) {
 		this.version = version;
 	}
-
+	
 	public String getImages() {
 		return images;
 	}
 	public void setImages(String images) {
 		this.images = images;
-	}
-	public String getISBN() {
-		return ISBN;
-	}
-	public void setISBN(String iSBN) {
-		ISBN = iSBN;
 	}
 	public String getDescription() {
 		return description;
