+++ src/cn/edu/zhku/jsj/domain/Food.java
@@ -1,4 +1,4 @@
-package cn.edu.zhku.jsj.daomain;
+package cn.edu.zhku.jsj.domain;
 
 public class Food {
 	private int food_id;
