+++ src/cn/edu/zhku/jsj/domain/PageBean.java
@@ -1,4 +1,4 @@
-package cn.edu.zhku.jsj.daomain;
+package cn.edu.zhku.jsj.domain;
 
 public class PageBean {
 	private int totalCount;//总记录数
