+++ src/cn/edu/zhku/jsj/domain/User.java
@@ -1,4 +1,4 @@
-package cn.edu.zhku.jsj.daomain;
+package cn.edu.zhku.jsj.domain;
 
 public class User {
 	private String user_id;
