+++ src/cn/edu/zhku/jsj/service/BusinessService.java
@@ -2,11 +2,11 @@ package cn.edu.zhku.jsj.service;
 
 import java.util.List;
 
-import cn.edu.zhku.jsj.daomain.Book;
-import cn.edu.zhku.jsj.daomain.Cloth;
-import cn.edu.zhku.jsj.daomain.Food;
-import cn.edu.zhku.jsj.daomain.Store;
-import cn.edu.zhku.jsj.daomain.User;
+import cn.edu.zhku.jsj.domain.Book;
+import cn.edu.zhku.jsj.domain.Cloth;
+import cn.edu.zhku.jsj.domain.Food;
+import cn.edu.zhku.jsj.domain.Store;
+import cn.edu.zhku.jsj.domain.User;
 import cn.edu.zhku.jsj.exception.UserexistException;
 
 public interface BusinessService {
