+++ src/cn/edu/zhku/jsj/service/impl/BusinessServiceImpl.java
@@ -10,11 +10,11 @@ import cn.edu.zhku.jsj.dao.OperatorDao;
 import cn.edu.zhku.jsj.dao.OrderDao;
 import cn.edu.zhku.jsj.dao.StoreDao;
 import cn.edu.zhku.jsj.dao.UserDao;
-import cn.edu.zhku.jsj.daomain.Book;
-import cn.edu.zhku.jsj.daomain.Cloth;
-import cn.edu.zhku.jsj.daomain.Food;
-import cn.edu.zhku.jsj.daomain.Store;
-import cn.edu.zhku.jsj.daomain.User;
+import cn.edu.zhku.jsj.domain.Book;
+import cn.edu.zhku.jsj.domain.Cloth;
+import cn.edu.zhku.jsj.domain.Food;
+import cn.edu.zhku.jsj.domain.Store;
+import cn.edu.zhku.jsj.domain.User;
 import cn.edu.zhku.jsj.exception.UserexistException;
 import cn.edu.zhku.jsj.factory.DaoFactory;
 import cn.edu.zhku.jsj.service.BusinessService;
