+++ src/cn/edu/zhku/jsj/web/shopkeeper/ModifyGoodServlet.java
@@ -10,9 +10,9 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
-import cn.edu.zhku.jsj.daomain.Book;
-import cn.edu.zhku.jsj.daomain.Cloth;
-import cn.edu.zhku.jsj.daomain.Food;
+import cn.edu.zhku.jsj.domain.Book;
+import cn.edu.zhku.jsj.domain.Cloth;
+import cn.edu.zhku.jsj.domain.Food;
 import cn.edu.zhku.jsj.formbean.BookFormBean;
 import cn.edu.zhku.jsj.formbean.ClothFormBean;
 import cn.edu.zhku.jsj.formbean.FoodFormBean;
