+++ src/cn/edu/zhku/jsj/web/utils/WebUtil.java
@@ -17,9 +17,9 @@ import org.apache.commons.fileupload.FileUploadBase.FileSizeLimitExceededExcepti
 import org.apache.commons.fileupload.disk.DiskFileItemFactory;
 import org.apache.commons.fileupload.servlet.ServletFileUpload;
 
-import cn.edu.zhku.jsj.daomain.Book;
-import cn.edu.zhku.jsj.daomain.Cloth;
-import cn.edu.zhku.jsj.daomain.Food;
+import cn.edu.zhku.jsj.domain.Book;
+import cn.edu.zhku.jsj.domain.Cloth;
+import cn.edu.zhku.jsj.domain.Food;
 import cn.edu.zhku.jsj.exception.NotChooseFileException;
 import cn.edu.zhku.jsj.exception.PictureTypeException;
 import cn.edu.zhku.jsj.formbean.BookFormBean;
