+++ src/junit/test/UserDaoTest.java
@@ -8,7 +8,7 @@ import java.sql.ResultSetMetaData;
 import org.junit.Test;
 
 import cn.edu.zhku.jsj.dao.impl.UserDaoImpl;
-import cn.edu.zhku.jsj.daomain.User;
+import cn.edu.zhku.jsj.domain.User;
 import cn.edu.zhku.jsj.exception.UserexistException;
 import cn.edu.zhku.jsj.service.BusinessService;
 import cn.edu.zhku.jsj.service.impl.BusinessServiceImpl;
